Update validation spec and remove unused types
Jira ticket: CAMS-584
